修复进程单实例线程退出问题

+++ include/ashe/cmd_line_parse.hpp
@@ -44,7 +44,7 @@ class ASHE_API CmdLineParser {
     bool hasVal(const std::wstring& key) const;
 
     std::wstring getVal(const std::wstring& key) const;
-    int getKeyCount() const;
+    size_t getKeyCount() const;
 
    private:
     void parse();
